+++ public/js/checkStatus.js
