+++ public/js/index.js
